Move the tests to standard wallaby findable dir. Closes #386
