+++ packages/interbit-ui-tools/package.json
@@ -9,7 +9,7 @@
     "build": "webpack --env dev && webpack --env build",
     "build:watch": "webpack --env dev && webpack --env build --watch",
     "dev": "webpack --progress --colors --watch --env dev",
-    "test": "mocha \"tests/**/*.test.js\""
+    "test": "mocha \"src/tests/**/*.test.js\""
   },
   "keywords": [
     "Interbit",
