+++ packages/interbit/package.json
@@ -4,12 +4,13 @@
   "description": "Requireable deployment utilities for interbit-cli",
   "main": "src/index.js",
   "scripts": {
-    "test": "mocha \"tests/**/*.test.js\""
+    "test": "mocha \"src/tests/**/*.test.js\""
   },
   "author": "BTL GROUP LTD",
   "license": "MIT",
   "dependencies": {
     "cheerio": "^1.0.0-rc.2",
+    "deep-diff": "^1.0.1",
     "fs-extra": "^5.0.0",
     "interbit-core": "0.8.2",
     "interbit-covenant-tools": "0.4.28",
