+++ packages/interbit/src/tests/chainManagement/watchChain.test.js
@@ -1,0 +1,3 @@
+describe('watchChain', () => {
+  it('errors if the genesis blocks have changed when chains already exist')
+})
