+++ packages/interbit/src/tests/updateIndexHtml.test.js
@@ -1,10 +1,7 @@
 const assert = require('assert')
 const cheerio = require('cheerio')
 
-const {
-  updateDom,
-  camelCaseToHyphenated
-} = require('../src/file/updateIndexHtml')
+const { updateDom, camelCaseToHyphenated } = require('../file/updateIndexHtml')
 
 const indexHtml = `<!DOCTYPE html>
 <html lang="en">
